Rename deriver_partial_eq_or_partial_ord.
